Refactor Footer
