--- src/components/layout/Footer/FooterLinks.vue
@@ -1,34 +1,0 @@
-<script setup>
-import { IconBrandGithub, IconWorld, IconBrandTelegram } from '@tabler/icons-vue';
-
-import { Button } from '@/components/ui/button';
-
-// Ссылки для отображения в подвале
-const links = [
-  {
-    name: 'GitHub',
-    url: 'https://github.com/itsTPM/foxford-tools',
-    icon: IconBrandGithub,
-  },
-  {
-    name: 'Telegram',
-    url: 'https://t.me/foxfordclips',
-    icon: IconBrandTelegram,
-  },
-  {
-    name: 'Сайт',
-    url: 'https://fox.itstpm.rocks',
-    icon: IconWorld,
-  },
-];
-</script>
-
-<template>
-  <div class="flex gap-3">
-    <Button size="icon" variant="outline" as-child v-for="link in links" :key="link.name">
-      <a :href="link.url" :title="link.name" :aria-label="link.name" target="_blank">
-        <component :is="link.icon" class="text-foreground w-6" stroke-width="1.5" aria-hidden="true" />
-      </a>
-    </Button>
-  </div>
-</template>
